+++ InstructionSets/x86/Descriptors.hpp
@@ -1,0 +1,26 @@
+//
+//  Descriptors.hpp
+//  Clock Signal
+//
+//  Created by Thomas Harte on 19/03/2025.
+//  Copyright © 2025 Thomas Harte. All rights reserved.
+//
+
+#pragma once
+
+namespace InstructionSet::x86 {
+
+enum class DescriptorTable {
+	Global, Local, Interrupt,
+};
+
+struct DescriptorTableLocation {
+	uint16_t limit;
+	uint32_t base;
+};
+
+struct Descriptor {
+	void set(uint64_t);
+};
+
+}
