+++ InstructionSets/x86/Implementation/PerformImplementation.hpp
@@ -20,8 +20,9 @@
 #include "ShiftRoll.hpp"
 #include "Stack.hpp"
 
-#include "InstructionSets/x86/Interrupts.hpp"
 #include "InstructionSets/x86/AccessType.hpp"
+#include "InstructionSets/x86/Descriptors.hpp"
+#include "InstructionSets/x86/Interrupts.hpp"
 
 //
 // Comments throughout headers above come from the 1997 edition of the
